cyber_prompt_injection/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
gunicorn==21.2.0 